Fix ID field doesn't display
