+++ packages/next-admin/src/types.ts
@@ -12,13 +12,13 @@ export type ObjectField<T extends ModelName> = Prisma.TypeMap["model"][T]["paylo
 
 export type Model<M extends ModelName, T extends never | number = never> = ScalarField<M> & {
   [P in keyof ObjectField<M>]:
-  ObjectField<M>[P] extends { scalars: infer S } ? T extends never ? S : T 
+  ObjectField<M>[P] extends { scalars: infer S } ? T extends never ? S : T
   : ObjectField<M>[P] extends { scalars: infer S } | null ? T extends never ? S | null : T | null
   : ObjectField<M>[P] extends { scalars: infer S }[] ? T extends never ? S[] : T[]
   : never;
 }
 
-export type ModelWithoutRelationships<M extends ModelName> = Model<M, number>; 
+export type ModelWithoutRelationships<M extends ModelName> = Model<M, number>;
 
 export type Field<P extends ModelName> = keyof Model<P>;
